requirements.txt
nvidia-riva-client
IPython
pandas